+++ src/app/manage/screen/screen-overview/clip-assigning-dialog/clip-assigning-dialog.module.ts
@@ -1,12 +1,14 @@
-import { NgModule } from "@angular/core";
-import { CommonModule } from "@angular/common";
-import { ClipAssigningDialogComponent } from "./clip-assigning-dialog/clip-assigning-dialog.component";
-import { MatListModule } from "@angular/material/list";
-import { MatIconModule } from "@angular/material/icon";
-import { MatDialogModule } from "@angular/material/dialog";
-import { MatButtonModule } from "@angular/material/button";
-import { PipesModule } from "../../../../core/pipes/pipes.module";
+import {NgModule} from "@angular/core";
+import {CommonModule} from "@angular/common";
+import {ClipAssigningDialogComponent} from "./clip-assigning-dialog/clip-assigning-dialog.component";
+import {MatListModule} from "@angular/material/list";
+import {MatIconModule} from "@angular/material/icon";
+import {MatDialogModule} from "@angular/material/dialog";
+import {MatButtonModule} from "@angular/material/button";
+import {PipesModule} from "../../../../core/pipes/pipes.module";
 import {MatChipsModule} from "@angular/material/chips";
+import {FilterModule} from "../../../../shared/components/filter/filter.module";
+
 // todo better name? :D
 
 @NgModule({
@@ -19,6 +21,7 @@ import {MatChipsModule} from "@angular/material/chips";
     MatDialogModule,
     MatButtonModule,
     MatChipsModule,
+    FilterModule,
   ]
 })
 export class ClipAssigningDialogModule {}
